# wanaku-router/wanaku-router-backend/src/main/webui/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    DataStore:
      type: object
      properties:
        labels:
          type: object
          additionalProperties:
            type: string
        id:
          type: string
        name:
          type: string
        data:
          type: string
    ForwardReference:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        namespace:
          type: string
...
    Namespace:
      type: object
      properties:
        labels:
          type: object
          additionalProperties:
            type: string
        id:
          type: string
        name:
          type: string
        path:
          type: string
...
    Param:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
...
    ResourceReference:
      type: object
      properties:
        labels:
          type: object
          additionalProperties:
            type: string
        id:
          type: string
        location:
          type: string
        type:
          type: string
        name:
          type: string
        description:
          type: string
        mimeType:
          type: string
        params:
          type: array
          items:
            $ref: "#/components/schemas/Param"
        configurationURI:
          type: string
        secretsURI:
          type: string
        namespace:
          type: string
...
    WanakuError:
      type: object
      properties:
        message:
          type: string
    WanakuResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data: {}
    WanakuResponseDataStore:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/DataStore"
    WanakuResponseForwardReference:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/ForwardReference"
    WanakuResponseInteger:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: integer
          format: int32
    WanakuResponseListDataStore:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: array
          items:
            $ref: "#/components/schemas/DataStore"
    WanakuResponseListForwardReference:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: array
          items:
            $ref: "#/components/schemas/ForwardReference"
    WanakuResponseListNamespace:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Namespace"
...
    WanakuResponseNamespace:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/Namespace"
...
    WanakuResponseResourceReference:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/WanakuError"
        data:
          $ref: "#/components/schemas/ResourceReference"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8543/realms/wanaku/.well-known/openid-configuration
      description: Authentication
paths:
...
  /api/v1/data-store:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataStore"
        required: true
      responses:
        "200":
          description: OK
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Data Stores Resource
    get:
      parameters:
      - name: labelFilter
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListDataStore"
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: List
      tags:
      - Data Stores Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataStore"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseDataStore"
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
        "400":
          description: Bad Request
      summary: Add
      tags:
      - Data Stores Resource
  /api/v1/data-store/by-label:
    delete:
      parameters:
      - name: labelExpression
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseInteger"
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: Remove If
      tags:
      - Data Stores Resource
  /api/v1/data-store/by-name:
    delete:
      parameters:
      - name: name
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: Remove By Name
      tags:
      - Data Stores Resource
    get:
      parameters:
      - name: name
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListDataStore"
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: Get By Name
      tags:
      - Data Stores Resource
  /api/v1/data-store/{id}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: Remove
      tags:
      - Data Stores Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseDataStore"
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: Get By Id
      tags:
      - Data Stores Resource
  /api/v1/forwards:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardReference"
        required: true
      responses:
        "200":
          description: OK
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Forwards Resource
    delete:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardReference"
        required: true
      responses:
        "200":
          description: OK
      summary: Remove Forward
      tags:
      - Forwards Resource
    get:
      parameters:
      - name: labelFilter
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListForwardReference"
      summary: List Forwards
      tags:
      - Forwards Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardReference"
        required: true
      responses:
        "200":
          description: OK
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
        "400":
          description: Bad Request
      summary: Add Forward
      tags:
      - Forwards Resource
  /api/v1/forwards/{name}:
    get:
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseForwardReference"
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: Get By Name
      tags:
      - Forwards Resource
...
  /api/v1/namespaces:
    get:
      parameters:
      - name: labelFilter
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseListNamespace"
      summary: List
      tags:
      - Namespaces Resource
  /api/v1/namespaces/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Namespace"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Namespaces Resource
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseNamespace"
      summary: Get By Id
      tags:
      - Namespaces Resource
...
  /api/v1/resources/{name}:
    get:
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponseResourceReference"
        "500":
          description: Wanaku error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WanakuResponse"
      summary: Get By Name
      tags:
      - Resources Resource
...
info:
  title: wanaku-router-backend API
  version: 0.0.9-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value